requirements.txt
python-telegram-bot
telethon
openai
telegraph
apscheduler
python-dotenv
